+++ e2e/src/helper.ts
@@ -219,16 +219,3 @@ export async function triggerCompletion(docUri: Uri, position: Position, trigger
 export async function triggerDefinition(docUri: Uri, position: Position): Promise<DefinitionLink[]> {
   return commands.executeCommand<DefinitionLink[]>('vscode.executeDefinitionProvider', docUri, position);
 }
-
-export async function assertDefinitions(docUri: Uri, position: Position, expectedDefinitions: DefinitionLink[]): Promise<void> {
-  const definitions = await triggerDefinition(docUri, position);
-
-  assertThat(definitions.length, expectedDefinitions.length);
-
-  for (let i = 0; i < definitions.length; i++) {
-    assertThat(definitions[i].originSelectionRange, expectedDefinitions[i].originSelectionRange);
-    assertThat(definitions[i].targetUri.path, expectedDefinitions[i].targetUri.path);
-    assertThat(definitions[i].targetRange, expectedDefinitions[i].targetRange);
-    assertThat(definitions[i].targetSelectionRange, expectedDefinitions[i].targetSelectionRange);
-  }
-}
